+++ packages/adminui-vue/components/UToolbar/UToolbar.vue
@@ -29,6 +29,7 @@
       <u-toolbar-button
         icon="fa fa-cog"
         color="secondary"
+        tooltip="allActions"
       />
 
       <template #dropdown>