+++ packages/adminui-pub/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+  - rename i18n key `vseDeystviya` -> `allActions` 
 
 ### Deprecated